requirements.txt
pyyaml
plyer